requirements.txt
pyyaml
plyer